+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## v1.1.0
+- New redis-atom should not overwrite existing key-value pair on backend.
+- Added no-value redis-atom arity.
+
 ## v1.0.0
 Support all clojure atom functionality